feat(address): fix issue with ui not implemented
