+++ lib/features/address/presentation/views/add_address.dart
@@ -4,8 +4,10 @@ import 'package:flower_app/features/address/presentation/views/select_location.d
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:geolocator/geolocator.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
 import '../../../../core/cubits/local_cubit/local_cubit.dart';
 import '../../data/models/cities_model.dart';
+import '../../data/models/search_result.dart';
 import '../../data/models/states_model.dart';
 import '../cubits/address_cubit/address_cubit.dart';
 import '../cubits/address_cubit/address_state.dart';
@@ -22,6 +24,7 @@ class _AddAddressScreenState extends State<AddAddressScreen>
   final TextEditingController addressController = TextEditingController();
   final TextEditingController recipientController = TextEditingController();
   final TextEditingController phoneController = TextEditingController();
+  GoogleMapController? mapController;
   bool inSettings = false;
 
   List<CitiesModel> cities = [];
@@ -38,14 +41,12 @@ class _AddAddressScreenState extends State<AddAddressScreen>
 
   Future<void> _loadInitialData() async {
     var cubit = AddressCubit.get(context);
-
     cubit.getCurrentAddress();
 
     cities = await cubit.getCities();
 
     if (cities.isNotEmpty) {
       selectedCityId = cities.first.id;
-
       await _loadStatesForCity(selectedCityId);
     }
   }
@@ -77,95 +78,106 @@ class _AddAddressScreenState extends State<AddAddressScreen>
   @override
   Widget build(BuildContext context) {
     var cubit = AddressCubit.get(context);
+
     return Scaffold(
       appBar: AppBar(title: const Text("Address")),
       body: SingleChildScrollView(
         padding: const EdgeInsets.all(16),
         child: BlocBuilder<AddressCubit, AddressState>(
           builder: (context, state) {
-            if (state is AddressSuccess && addressController.text.isEmpty) {
+            print(state);
+            if (state is AddressSuccess) {
               addressController.text = state.searchResult.address;
             }
 
             return Column(
               children: [
-                GestureDetector(
-                  onTap: () async {
-                    if (state is AddressAccessDenied) {
-                      showDialog(
-                        context: context,
-                        builder: (context) {
-                          return AlertDialog(
-                            backgroundColor: Colors.white,
-                            title: const Text("Location Access Denied"),
-                            actions: [
-                              TextButton(
-                                onPressed: () {
-                                  Navigator.pop(context);
-                                  openLocationSettings();
-                                },
-                                child: const Text("Allow"),
-                              ),
-                              TextButton(
-                                onPressed: () {
-                                  Navigator.pop(context);
-                                },
-                                child: Text(
-                                  "Cancel",
-                                  style: TextStyle(
-                                    color: ColorManager.appColor,
-                                  ),
-                                ),
-                              ),
-                            ],
-                          );
-                        },
-                      );
-                    }
-                    if (state is AddressSuccess) {
-                      var result = await Navigator.push(
-                        context,
+                Stack(
+                  children: [
+                    Container(
+                      height: 150,
+                      width: double.infinity,
+                      margin: const EdgeInsets.only(bottom: 16),
+                      decoration: BoxDecoration(
+                        color: Colors.grey[200],
+                        border: Border.all(color: Colors.grey),
+                        borderRadius: BorderRadius.circular(12),
+                      ),
+                      child: buildMap(state),
+                    ),
+                    Positioned.fill(
+                      child: Material(
+                        color: Colors.transparent,
+                        child: GestureDetector(
+                          onTap: () async {
+                            print("Map tapped. Current state: $state");
 
-                        MaterialPageRoute(
-                          builder:
-                              (context) => SelectLocationScreen(
-                                initialPosition: state.searchResult.location,
+                            if (state is AddressAccessDenied) {
+                              showDialog(
+                                context: context,
+                                builder: (context) {
+                                  return AlertDialog(
+                                    backgroundColor: Colors.white,
+                                    title: const Text("Location Access Denied"),
+                                    actions: [
+                                      TextButton(
+                                        onPressed: () {
+                                          Navigator.pop(context);
+                                          openLocationSettings();
+                                        },
+                                        child: const Text("Allow"),
+                                      ),
+                                      TextButton(
+                                        onPressed: () {
+                                          Navigator.pop(context);
+                                        },
+                                        child: Text(
+                                          "Cancel",
+                                          style: TextStyle(
+                                            color: ColorManager.appColor,
+                                          ),
+                                        ),
+                                      ),
+                                    ],
+                                  );
+                                },
+                              );
+                              return;
+                            }
+
+                            if (state is! AddressSuccess) {
+                              await cubit.getCurrentAddress();
+                              return;
+                            }
+
+                            final result = await Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder:
+                                    (context) => SelectLocationScreen(
+                                      initialPosition:
+                                          (state).searchResult.location,
+                                    ),
                               ),
-                          settings: RouteSettings(
-                            arguments: state.searchResult.location,
-                          ),
-                        ),
-                      );
-                      if (result != null) {
-                        addressController.clear();
-                        addressController.text = result['address'];
-                      }
-                    }
-                  },
-                  child: Stack(
-                    children: [
-                      Container(
-                        height: 150,
-                        width: double.infinity,
-                        margin: const EdgeInsets.only(bottom: 16),
-                        decoration: BoxDecoration(
-                          color: Colors.grey[200],
-                          border: Border.all(color: Colors.grey),
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        child: cubit.buildMap(state),
-                      ),
-                      Positioned.fill(
-                        child: Material(
-                          color: Colors.transparent,
-                          child: InkWell(
-                            borderRadius: BorderRadius.circular(12),
-                            splashColor: Colors.pink,
-                          ),
+                            );
+
+                            if (result != null) {
+                              cubit.updateAddress(
+                                SearchResult(
+                                  address: result['address'],
+                                  location: result['location'],
+                                ),
+                              );
+
+                              mapController?.animateCamera(
+                                CameraUpdate.newLatLng(result['location']),
+                              );
+                            }
+                          },
                         ),
                       ),
-                    ],
-                  ),
+                    ),
+                  ],
                 ),
                 TextField(
                   controller: addressController,
@@ -224,12 +236,14 @@ class _AddAddressScreenState extends State<AddAddressScreen>
                             states.map((state) {
                               return DropdownMenuItem<String>(
                                 value: state.id,
-                                child: Text(  LocalizationCubit.get(
-                                  context,
-                                ).state.language ==
-                                    'en'
-                                    ? state.stateNameEn
-                                    : state.stateNameAr,),
+                                child: Text(
+                                  LocalizationCubit.get(
+                                            context,
+                                          ).state.language ==
+                                          'en'
+                                      ? state.stateNameEn
+                                      : state.stateNameAr,
+                                ),
                               );
                             }).toList(),
                         onChanged: (String? stateId) {
@@ -242,40 +256,64 @@ class _AddAddressScreenState extends State<AddAddressScreen>
                   ],
                 ),
                 const SizedBox(height: 20),
-                // ElevatedButton(
-                //   style: ElevatedButton.styleFrom(
-                //     minimumSize: const Size(double.infinity, 50),
-                //     backgroundColor: Colors.grey[700],
-                //     shape: RoundedRectangleBorder(
-                //       borderRadius: BorderRadius.circular(12),
-                //     ),
-                //   ),
-                //   onPressed: () {
-                //     String cityName = cities.firstWhere(
-                //             (city) => city.id == selectedCityId,
-                //         orElse: () => CitiesModel(id: '', name: 'Unknown')
-                //     ).name;
-                //
-                //     String stateName = states.firstWhere(
-                //             (state) => state.id == selectedStateId,
-                //         orElse: () => StatesModel(id: '', name: 'Unknown')
-                //     ).name;
-                //
-                //     print("Saved Address: ${addressController.text}");
-                //     print("City: $cityName, Area: $stateName");
-                //     print("Recipient: ${recipientController.text}");
-                //     print("Phone: ${phoneController.text}");
-                //   },
-                //   child: const Text(
-                //     "Save address",
-                //     style: TextStyle(color: Colors.white),
-                //   ),
-                // ),
               ],
             );
           },
         ),
       ),
     );
+  }
+
+  Widget buildMap(AddressState state) {
+    if (state is AddressSuccess) {
+      return ClipRRect(
+        borderRadius: BorderRadius.circular(12),
+        child: GoogleMap(
+          onMapCreated: (controller) {
+            mapController = controller;
+          },
+          initialCameraPosition: CameraPosition(
+            target: state.searchResult.location,
+            zoom: 15,
+          ),
+          markers: {
+            Marker(
+              markerId: const MarkerId("current"),
+              position: state.searchResult.location,
+              infoWindow: InfoWindow(title: state.searchResult.address),
+            ),
+          },
+          mapToolbarEnabled: false,
+          zoomControlsEnabled: false,
+          myLocationButtonEnabled: false,
+          rotateGesturesEnabled: false,
+          scrollGesturesEnabled: false,
+          zoomGesturesEnabled: false,
+          tiltGesturesEnabled: false,
+          liteModeEnabled: true,
+        ),
+      );
+    } else if (state is AddressError) {
+      return Center(child: Text(state.message));
+    } else if (state is AddressLoading) {
+      return Center(
+        child: CircularProgressIndicator(color: ColorManager.appColor),
+      );
+    } else if (state is AddressAccessDenied) {
+      return Center(
+        child: TextButton(
+          onPressed: openLocationSettings,
+          child: Text(
+            "Allow access to location",
+            style: TextStyle(
+              color: ColorManager.appColor,
+              fontSize: 20,
+              fontWeight: FontWeight.w700,
+            ),
+          ),
+        ),
+      );
+    }
+    return Container();
   }
 }
